requirements.txt
Pillow
PyOpenGL
pygame
pandas
seaborn
scikit-learn
joblib
opencv-python
pyyaml
onnxruntime
vidgear
numba
